move to slot
