+++ src/panels/energy/ha-panel-energy.ts
@@ -7,8 +7,10 @@ import {
   TemplateResult,
   nothing,
 } from "lit";
+import { mdiPencil } from "@mdi/js";
 import { customElement, property, state } from "lit/decorators";
 import "../../components/ha-menu-button";
+import "../../components/ha-list-item";
 import { LovelaceConfig } from "../../data/lovelace";
 import { haStyle } from "../../resources/styles";
 import { HomeAssistant } from "../../types";
@@ -16,6 +18,7 @@ import "../lovelace/components/hui-energy-period-selector";
 import { Lovelace } from "../lovelace/types";
 import "../lovelace/views/hui-view";
 import "../../components/ha-top-app-bar-fixed";
+import { navigate } from "../../common/navigate";
 
 const ENERGY_LOVELACE_CONFIG: LovelaceConfig = {
   views: [
@@ -70,9 +73,24 @@ class PanelEnergy extends LitElement {
 
           <hui-energy-period-selector
             .hass=${this.hass}
-            showConfig
             collectionKey="energy_dashboard"
-          ></hui-energy-period-selector>
+          >
+            ${this.hass.user?.is_admin
+              ? html`
+                  <ha-list-item
+                    slot="overflow-menu"
+                    graphic="icon"
+                    @request-selected=${this._navigateConfig}
+                  >
+                    <ha-svg-icon slot="graphic" .path=${mdiPencil}>
+                    </ha-svg-icon>
+                    ${this.hass!.localize(
+                      "ui.panel.lovelace.menu.configure_ui"
+                    )}
+                  </ha-list-item>
+                `
+              : nothing}
+          </hui-energy-period-selector>
         </div>
       </div>
       <hui-view
@@ -99,6 +117,11 @@ class PanelEnergy extends LitElement {
       deleteConfig: async () => undefined,
       setEditMode: () => undefined,
     };
+  }
+
+  private _navigateConfig(ev) {
+    ev.stopPropagation();
+    navigate("/config/energy?historyBack=1");
   }
 
   private _reloadView() {
